feat : 유저 업데이트 로직에 추가된 필드들도 수정하게끔
