+++ src/main/java/com/runningRank/runningRank/auth/dto/UserUpdateRequest.java
@@ -8,14 +8,15 @@ import java.time.LocalDate;
 public class UserUpdateRequest {
     private String profileImageUrl;
     private String name;
-    private boolean isNameVisible;
     private LocalDate birthDate;
     private String gender;
     private String universityEmail;
     private String universityName;
     private String studentNumber;
-    private boolean isStudentNumberVisible;
     private String major;
-    private boolean isMajorVisible;
     private boolean isChangeUniversity;
+    private boolean isNameVisible;
+    private boolean isStudentNumberVisible;
+    private boolean isMajorVisible;
+    private String graduationStatus;
 }
